Move all directory scanning to dir sub-package (#1)
- Mode all glob related functions to sub-package dir
- Add support for symlinks travesal
- Add fileutil.Touch() that touches one of more files

- TODO: Need more coverage of error cases in atomic Read / Write
- TODO: Need more testing

+++ atomic_test.go
@@ -3,7 +3,10 @@ package fileutil_test
 import (
 	"encoding/json"
 	"io"
+	"io/ioutil"
 	"os"
+	"path/filepath"
+	"sync"
 	"testing"
 
 	"github.com/maargenton/fileutil"
@@ -12,16 +15,21 @@ import (
 )
 
 func TestOpenTemp(t *testing.T) {
-	assert := asserter.New(t)
-	assert.That(nil, p.IsNil())
+	assert := asserter.New(t, asserter.AbortOnError())
+	dir, err := ioutil.TempDir(".", "testdata-")
+	assert.That(err, p.IsNoError())
+	defer os.RemoveAll(dir) // clean up
 
-	f, err := fileutil.OpenTemp("testdata/file.txt", "tmp")
-	defer os.Remove(f.Name())
+	filename := filepath.Join(dir, "file.txt")
+	f, err := fileutil.OpenTemp(filename, "tmp")
+	if f != nil {
+		defer os.Remove(f.Name())
+	}
 	defer f.Close()
 
 	assert.That(err, p.IsNoError())
 	assert.That(f, p.IsNotNil())
-	assert.That(f.Name(), p.StartsWith("testdata/file-tmp"))
+	assert.That(f.Name(), p.StartsWith(filepath.Join(dir, "file")))
 }
 
 // ---------------------------------------------------------------------------
@@ -30,27 +38,59 @@ type Content struct {
 	Seq int `json:"seq,omitempty"`
 }
 
-func TestWriteFile(t *testing.T) {
+func TestReadWriteFile(t *testing.T) {
 	assert := asserter.New(t)
-	assert.That(nil, p.IsNil())
+	dir, err := ioutil.TempDir(".", "testdata-")
+	assert.That(err, p.IsNoError())
+	defer os.RemoveAll(dir) // clean up
 
+	// Write file
 	var content = &Content{Seq: 125}
-	err := fileutil.WriteFile("testdata/file.txt", func(w io.Writer) error {
+	filename := filepath.Join(dir, "file.txt")
+	err = fileutil.WriteFile(filename, func(w io.Writer) error {
 		return json.NewEncoder(w).Encode(content)
 	})
-
 	assert.That(err, p.IsNoError())
-}
 
-func TestReadFile(t *testing.T) {
-	assert := asserter.New(t)
-	assert.That(nil, p.IsNil())
-
-	var content = &Content{}
-	err := fileutil.ReadFile("testdata/file.txt", func(r io.Reader) error {
+	// Read file
+	content = &Content{}
+	err = fileutil.ReadFile(filename, func(r io.Reader) error {
 		return json.NewDecoder(r).Decode(content)
 	})
 
 	assert.That(err, p.IsNoError())
 	assert.That(content.Seq, p.Eq(125))
+}
+
+func TestWriteFileIsAtomic(t *testing.T) {
+	assert := asserter.New(t)
+	dir, err := ioutil.TempDir(".", "testdata-")
+	assert.That(err, p.IsNoError())
+	defer os.RemoveAll(dir) // clean up
+
+	filename := filepath.Join(dir, "file.txt")
+	var wg sync.WaitGroup
+	for i := 0; i < 10; i++ {
+		i := i
+		wg.Add(1)
+
+		go func() {
+			content := &Content{Seq: i}
+			err := fileutil.WriteFile(filename, func(w io.Writer) error {
+				return json.NewEncoder(w).Encode(content)
+			})
+			assert.That(err, p.IsNoError())
+			wg.Done()
+		}()
+	}
+
+	wg.Wait()
+	content := &Content{}
+	err = fileutil.ReadFile(filename, func(r io.Reader) error {
+		return json.NewDecoder(r).Decode(content)
+	})
+
+	assert.That(err, p.IsNoError())
+	assert.That(content.Seq, p.Ge(0))
+	assert.That(content.Seq, p.Lt(10))
 }
